Use Long to represent Content-Length size (#357)
* Use Long for a Content-Length header size.
* Ensure content length is non-negative number.

+++ components/api/src/main/java/com/hotels/styx/api/HttpMessage.java
@@ -84,8 +84,8 @@ interface HttpMessage {
      *
      * @return the content-length if present
      */
-    default Optional<Integer> contentLength() {
-        return header(CONTENT_LENGTH).map(Integer::valueOf);
+    default Optional<Long> contentLength() {
+        return header(CONTENT_LENGTH).map(Long::valueOf);
     }
 
     /**
